+++ API/Controllers/CategoryController.cs
@@ -7,6 +7,7 @@ using API.Errors;
 using AutoMapper;
 using Core.Entities;
 using Core.Interfaces;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace API.Controllers
@@ -28,8 +29,7 @@ namespace API.Controllers
             try
             {
                 var categoriesList = await _categoryService.ListAllCategoryAsync();
-                // var categoryListDto = _mapper.Map<CategoryDto>(categoriesList);
-
+       
                 return Ok(categoriesList);
             }
             catch (Exception ex)
@@ -58,6 +58,7 @@ namespace API.Controllers
 
         }
 
+        [Authorize]
         [HttpPost]
         public async Task<ActionResult<AgentsDto>> CreateCategoryAsync(CategoryDto category)
         {
@@ -75,11 +76,19 @@ namespace API.Controllers
 
         }
 
+
+        [Authorize]
         [HttpPut]
-        public async Task<ActionResult<AgentsDto>> UpdateAgentAsync(CategoryDto category)
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(typeof(ApiResponse), StatusCodes.Status404NotFound)]
+        public async Task<ActionResult<AgentsDto>> UpdateCategoryAsync(CategoryDto category)
         {
             try
             {
+                var singleCategory = await _categoryService.GetCategoryByIdAsync(category.Id);
+                var categoryDto = _mapper.Map<CategoryDto>(singleCategory);
+
+                if (categoryDto == null) return NotFound(new ApiResponse(404));
                 var fromDto = _mapper.Map<Category>(category);
                 var result = await _categoryService.UpdateCategory(fromDto);
                 return Ok(category);
@@ -91,11 +100,19 @@ namespace API.Controllers
 
         }
 
+
+        [Authorize]
         [HttpDelete("{id}")]
-        public async Task<ActionResult<bool>> DeleteAgent(int id)
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(typeof(ApiResponse), StatusCodes.Status404NotFound)]
+        public async Task<ActionResult<bool>> DeleteCategory(int id)
         {
             try
             {
+                var singleCategory = await _categoryService.GetCategoryByIdAsync(id);
+                var categoryDto = _mapper.Map<CategoryDto>(singleCategory);
+                if (categoryDto == null) return NotFound(new ApiResponse(404));
+
                 var deletedCategory = await _categoryService.DeleteCategory(id);
                 if (deletedCategory == null) return false;
                 return true;
